+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 0.1.0 (1/27/2013)
 ==================
 
-* Initial public release - [@dblock](https://github.com/dblock).
+* Initial public release with support for `mongodump` - [@dblock](https://github.com/dblock).